fix: faulty navigation path when navbar item is clicked more than once and autoselect text box for rfid
